+++ src/index.js
@@ -12,3 +12,4 @@ export * from './layouts/application/index.js';
 export * from './extended-search/index.js';
 export * from './wizard/index.js';
 export * from './menu/index.js';
+export * from './constants/index.js';